Use long in timestamp

+++ API Models/LolItemSetsItemSets.cs
@@ -4,6 +4,6 @@ namespace LCU.NET
     {
         public int accountId { get; set; }
         public LolItemSetsItemSet[] itemSets { get; set; }
-        public int timestamp { get; set; }
+        public long timestamp { get; set; }
     }
 }
